// web/src/app/store/model/store.model.ts
export class Store {
  storeAddressNumber: string;
  storeAddress: string;
  storeDescription: string;
  storeFloor: string;
  storeMail: string;
  storeName: string;
  storePhone: string;
  storePosition: string;
  userName: string;
  facebookLink: string;
  instagramLink: String;
  twitterLink: String;
}